Rename pydantic to schemas
